Action system! Will be used to make an undo/redo system
Ok, this is a huge commit with only half of the stuff working as
intended. But I just wanted to push at least something because I think I finaly found the way.

It was really hard to find a good architechture, you can all
tell by the time since the last commit ...